The Watering Can is now fully functional and waters the plant

+++ Assets/Scripts/Classes/PlantClass.cs
@@ -226,7 +226,7 @@ namespace Classes.Plants
     {
         return waterMaxRange;
     }
-    public void setSunlightAmount(int amount)
+    public void setSunlightAmount(double amount)
     {
         sunlightAmount = amount;
     }
